Implemented and fixed subcomponent loading
Fixed the loading of JS components, including web components. Fixed the loading of CSS and SCSS stylesheets. Created functions to automatically add all scss and js files for that respective page in the entry points to then be loaded with the page. Added loading for images and vectors. Improved loading for HTML pages. Implemented loading of JS components with babel loader.