Optimised built artifacts and updated artifact naming
Instead of archiving the entire repo as an artifact for builds, only archive the build repository to allow for optimisation in the project setup workflow. As such added a checkout in the deployment workflow to prevent issues with the gh pages deployment. Also updated the artifact naming by changing non-built artifacts to be named 'source-[sha]' instead of 'no-build-[sha]'